// deskty/src/app/models/profile.model.ts
export class profileModels {
  id?: any;
  profileId: any;
  firstName: any;
  lastName: any;
  accountCount: any;
  profileType: any;
  strictMode!: boolean;
  holidayMode!: boolean;
  profilePassword: any;
  profileUrl: any;
  username: any;
  nudge!: boolean;
  ssrequest!: boolean;
  startSessionDate: any;
  endSessionDate: any;
}